+++ viscy/light/engine.py
@@ -29,7 +29,7 @@ from viscy.data.hcs import Sample
 from viscy.evaluation.evaluation_metrics import mean_average_precision, ms_ssim_25d
 from viscy.unet.networks.fcmae import FullyConvolutionalMAE
 from viscy.unet.networks.Unet2D import Unet2d
-from viscy.unet.networks.Unet22D import Unet22d
+from viscy.unet.networks.unext2 import Unet22d
 from viscy.unet.networks.Unet25D import Unet25d
 
 try:
